// client/src/spikes/components/SpikeStatusChip.d.ts
import React from 'react';
import type { SpikeStatus } from '../types';
interface SpikeStatusChipProps {
    status: SpikeStatus;
    label?: string;
    className?: string;
}
export declare const SpikeStatusChip: React.FC<SpikeStatusChipProps>;
export default SpikeStatusChip;
//# sourceMappingURL=SpikeStatusChip.d.ts.map